+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.42b4"  # see https://peps.python.org/pep-0440/
-    AppDateStr = "2026-02-04"
+    AppVersionStr = "0.43b1"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2026-02-13"